fix: ST_NumInteriorRing failing with wrong return type

+++ velox/functions/prestosql/tests/GeometryFunctionsTest.cpp
@@ -2008,8 +2008,8 @@ TEST_F(GeometryFunctionsTest, testStInteriorRingN) {
 TEST_F(GeometryFunctionsTest, testStNumInteriorRing) {
   const auto testStNumInteriorRingFunc =
       [&](const std::optional<std::string>& wkt,
-          const std::optional<int32_t>& expected) {
-        std::optional<int32_t> result = evaluateOnce<int32_t>(
+          const std::optional<int64_t>& expected) {
+        std::optional<int64_t> result = evaluateOnce<int64_t>(
             "ST_NumInteriorRing(ST_GeometryFromText(c0))", wkt);
 
         if (expected.has_value()) {
